requirements.txt
streamlit
requests
beautifulsoup4
duckduckgo-search
openai
python-dotenv